=General *Fixed an Issue where hatch a pet could disconnect you.
*Added a new value in conf/mob.conf to Enable that mobs'll teleport when standing in Dimensional Doors portals

*Minor fix to give support to the new options in clones.

*Merged missing changes from 3ceam rev. 10 - 31

=Database
*updated skill_unit_db.txt


=Skills
*AB_ADORAMUS
*AB_RENOVATIO
*NC_ANALYZE
*NC_BOOSTKNUCKLE
*SR_POWERVELOCITY
*WL_SOULEXPANSION
-Added pre-support for these skills.

*RA_AIMEDBOLT
*RK_PHANTOMTHRUST
*WL_CHAINLIGHTNING
-Some more work on these skills.

*SR_GATEOFHELL
-Added full support for this skill, however the damage formular needs to be updated.


=Status
*SC_DIMENSIONDOOR
-adjusted its behavior to the official. This status don't work on mobs.

+++ trunk/src/map/mob.h
@@ -103,7 +103,7 @@ struct mob_db {
 	} mvpitem[3];
 	struct status_data status;
 	struct view_data vd;
-	short option;
+	int option; // Changed to int to give support to the new options. [pakpil] // 3ceam v1.
 	int summonper[MAX_RANDOMMONSTER];
 	int maxskill;
 	struct mob_skill skill[MAX_MOBSKILL];
